+++ src/screens/discover/DiscoverDetailScreen.tsx
