+++ src/screens/discover/index.ts
@@ -1,0 +1,4 @@
+import DiscoverScreen from './DiscoverScreen';
+import DiscoverDetailScreen from './DiscoverDetailScreen';
+
+export { DiscoverScreen, DiscoverDetailScreen };
